backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
firebase-admin==6.2.0
google-generativeai==0.3.1
pytesseract==0.3.10
pdfplumber==0.10.3
Pillow==10.1.0
python-multipart==0.0.6
apscheduler==3.10.4
httpx==0.25.2
requests==2.31.0
pydantic==2.5.0
